Remove redundant additional dependency
pydantic-settings pulls in pydantic, so there is no need to list the
latter as a dependency to the mypy pre-commit linter.

This also puts the mypy pre-commit linter dependencies in sync with
thpse listed in the dev section of extras_require in setup.py.  The
intent is that these two lists of dependencies always be in sync, so
this is good.